Move salesforcedx-vscode-lwc tests to vscode-integration dir
